example/eeprom: fix getopt, verbose
Optional argument, two colons - "::", doesn't work on my Linux
(Ubuntu 16.04, 64-bit). Remove it for the two reasons:
1. It's GNU extension and not POSIX standard.
2. I'm using GNU libc, but it doesn't work.

Separate the large_chip to use another option - 'l'

verbose option mentioned in the help menu, but not working.
Use 'b' option for it

Re-ordering the options so they are in the same order.